app/requirements/base.txt
autopep8==1.6.0
boto3==1.24.32
botocore==1.27.32
dj-database-url==0.5.0
Django==4.1
django-widget-tweaks==1.4.12
django-cors-headers==3.13.0
django-filter==22.1
django-storages==1.12.3
docutils==0.19
packaging==21.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
sendgrid==6.9.7
whitenoise==6.2.0
email-validator==1.2.1
channels==3.0.5
channels-redis==3.4.1
pytest-django==4.5.2
celery==5.2.7
redis==4.3.4
django-redis==5.2.0
flower==1.1.0
locust==2.10.1
pylint==2.14.5
pyfcm==1.5.4
pyotp==2.6.0
sentry-sdk==1.7.2
django-import-export==2.8.0
twilio==7.11.0
djangorestframework==3.13.1
Markdown==3.4.1
MarkupSafe==2.1.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
Faker==13.15.0
selenium==4.3.0
psutil==5.9.1
django-debug-toolbar==3.5.0
gunicorn==20.1.0
